+++ src/rcc/enable.rs
@@ -2,12 +2,12 @@ use super::*;
 use crate::bb;
 
 macro_rules! bus_enable {
-    ($PER:ident => ($busenr:ident, $enbit:literal)) => {
+    ($PER:ident => ($busX:ty, $bit:literal)) => {
         impl Enable for crate::pac::$PER {
             #[inline(always)]
             fn enable(rcc: &RccRB) {
                 unsafe {
-                    bb::set(&rcc.$busenr, $enbit);
+                    bb::set(Self::Bus::enr(rcc), $bit);
                 }
                 // Stall the pipeline to work around erratum 2.1.13 (DM00037591)
                 cortex_m::asm::dsb();
@@ -15,19 +15,19 @@ macro_rules! bus_enable {
             #[inline(always)]
             fn disable(rcc: &RccRB) {
                 unsafe {
-                    bb::clear(&rcc.$busenr, $enbit);
+                    bb::clear(Self::Bus::enr(rcc), $bit);
                 }
             }
         }
     };
 }
 macro_rules! bus_lpenable {
-    ($PER:ident => ($buslpenr:ident, $lpenbit:literal)) => {
+    ($PER:ident => ($busX:ty, $bit:literal)) => {
         impl LPEnable for crate::pac::$PER {
             #[inline(always)]
             fn low_power_enable(rcc: &RccRB) {
                 unsafe {
-                    bb::set(&rcc.$buslpenr, $lpenbit);
+                    bb::set(Self::Bus::lpenr(rcc), $bit);
                 }
                 // Stall the pipeline to work around erratum 2.1.13 (DM00037591)
                 cortex_m::asm::dsb();
@@ -35,20 +35,20 @@ macro_rules! bus_lpenable {
             #[inline(always)]
             fn low_power_disable(rcc: &RccRB) {
                 unsafe {
-                    bb::clear(&rcc.$buslpenr, $lpenbit);
+                    bb::clear(Self::Bus::lpenr(rcc), $bit);
                 }
             }
         }
     };
 }
 macro_rules! bus_reset {
-    ($PER:ident => ($busrstr:ident, $resetbit:literal)) => {
+    ($PER:ident => ($busX:ty, $bit:literal)) => {
         impl Reset for crate::pac::$PER {
             #[inline(always)]
             fn reset(rcc: &RccRB) {
                 unsafe {
-                    bb::set(&rcc.$busrstr, $resetbit);
-                    bb::clear(&rcc.$busrstr, $resetbit);
+                    bb::set(Self::Bus::rstr(rcc), $bit);
+                    bb::clear(Self::Bus::rstr(rcc), $bit);
                 }
             }
         }
@@ -56,74 +56,81 @@ macro_rules! bus_reset {
 }
 
 macro_rules! bus {
-    ($($PER:ident => ($busenr:ident, $buslpenr:ident, $busrstr:ident, $bit:literal),)+) => {
+    ($($PER:ident => ($busX:ty, $bit:literal),)+) => {
         $(
-            impl private::Sealed for crate::pac::$PER {}
-            bus_enable!($PER => ($busenr, $bit));
-            bus_lpenable!($PER => ($buslpenr, $bit));
-            bus_reset!($PER => ($busrstr, $bit));
+            impl Sealed for crate::pac::$PER {}
+            impl RccBus for crate::pac::$PER {
+                type Bus = $busX;
+            }
+            bus_enable!($PER => ($busX, $bit));
+            bus_lpenable!($PER => ($busX, $bit));
+            bus_reset!($PER => ($busX, $bit));
         )+
     }
 }
 
 bus! {
-    CRC => (ahb1enr, ahb1lpenr, ahb1rstr, 12),
-    DMA1 => (ahb1enr, ahb1lpenr, ahb1rstr, 21),
-    DMA2 => (ahb1enr, ahb1lpenr, ahb1rstr, 22),
+    CRC => (AHB1, 12),
+    DMA1 => (AHB1, 21),
+    DMA2 => (AHB1, 22),
 }
 
 bus! {
-    GPIOA => (ahb1enr, ahb1lpenr, ahb1rstr, 0),
-    GPIOB => (ahb1enr, ahb1lpenr, ahb1rstr, 1),
-    GPIOC => (ahb1enr, ahb1lpenr, ahb1rstr, 2),
-    GPIOH => (ahb1enr, ahb1lpenr, ahb1rstr, 7),
+    GPIOA => (AHB1, 0),
+    GPIOB => (AHB1, 1),
+    GPIOC => (AHB1, 2),
+    GPIOH => (AHB1, 7),
 }
 
 #[cfg(any(feature = "gpiod", feature = "gpioe"))]
 bus! {
-    GPIOD => (ahb1enr, ahb1lpenr, ahb1rstr, 3),
-    GPIOE => (ahb1enr, ahb1lpenr, ahb1rstr, 4),
+    GPIOD => (AHB1, 3),
+    GPIOE => (AHB1, 4),
 }
 #[cfg(any(feature = "gpiof", feature = "gpiog"))]
 bus! {
-    GPIOF => (ahb1enr, ahb1lpenr, ahb1rstr, 5),
-    GPIOG => (ahb1enr, ahb1lpenr, ahb1rstr, 6),
+    GPIOF => (AHB1, 5),
+    GPIOG => (AHB1, 6),
 }
 
 #[cfg(feature = "gpioi")]
 bus! {
-    GPIOI => (ahb1enr, ahb1lpenr, ahb1rstr, 8),
+    GPIOI => (AHB1, 8),
 }
 
 #[cfg(any(feature = "gpioj", feature = "gpiok"))]
 bus! {
-    GPIOJ => (ahb1enr, ahb1lpenr, ahb1rstr, 9),
-    GPIOK => (ahb1enr, ahb1lpenr, ahb1rstr, 10),
+    GPIOJ => (AHB1, 9),
+    GPIOK => (AHB1, 10),
 }
 
 #[cfg(feature = "rng")]
 bus! {
-    RNG => (ahb2enr, ahb2lpenr, ahb2rstr, 6),
+    RNG => (AHB2, 6),
 }
 
 #[cfg(feature = "otg-fs")]
 bus! {
-    OTG_FS_GLOBAL => (ahb2enr, ahb2lpenr, ahb2rstr, 7),
+    OTG_FS_GLOBAL => (AHB2, 7),
 }
 
 #[cfg(feature = "otg-hs")]
 bus! {
-    OTG_HS_GLOBAL => (ahb1enr, ahb1lpenr, ahb1rstr, 29),
+    OTG_HS_GLOBAL => (AHB1, 29),
 }
 
 #[cfg(feature = "fmc")]
 bus! {
-    FMC => (ahb3enr, ahb3lpenr, ahb3rstr, 0),
+    FMC => (AHB3, 0),
 }
 
 // TODO: fix absent ahb3lpenr
 #[cfg(feature = "fsmc")]
-impl private::Sealed for crate::pac::FSMC {}
+impl Sealed for crate::pac::FSMC {}
+#[cfg(feature = "fsmc")]
+impl RccBus for crate::pac::FSMC {
+    type Bus = AHB3;
+}
 #[cfg(feature = "fsmc")]
 #[cfg(any(feature = "stm32f427", feature = "stm32f437"))]
 bus_enable!(FSMC => (ahb3enr, 0));
@@ -138,119 +145,127 @@ bus_reset!(FSMC => (ahb3rstr, 0));
 bus_reset!(FSMC => (ahb3rstr, 0));
 
 bus! {
-    PWR => (apb1enr, apb1lpenr, apb1rstr, 28),
+    PWR => (APB1, 28),
 }
 
 bus! {
-    SPI1 => (apb2enr, apb2lpenr, apb2rstr, 12),
-    SPI2 => (apb1enr, apb1lpenr, apb1rstr, 14),
+    SPI1 => (APB2, 12),
+    SPI2 => (APB1, 14),
 }
 #[cfg(feature = "spi3")]
 bus! {
-    SPI3 => (apb1enr, apb1lpenr, apb1rstr, 15),
+    SPI3 => (APB1, 15),
 }
 
 #[cfg(feature = "spi4")]
 bus! {
-    SPI4 => (apb2enr, apb2lpenr, apb2rstr, 13),
+    SPI4 => (APB2, 13),
 }
 
 #[cfg(feature = "spi5")]
 bus! {
-    SPI5 => (apb2enr, apb2lpenr, apb2rstr, 20),
+    SPI5 => (APB2, 20),
 }
 
 #[cfg(feature = "spi6")]
 bus! {
-    SPI6 => (apb2enr, apb2lpenr, apb2rstr, 21),
+    SPI6 => (APB2, 21),
 }
 
 bus! {
-    I2C1 => (apb1enr, apb1lpenr, apb1rstr, 21),
-    I2C2 => (apb1enr, apb1lpenr, apb1rstr, 22),
+    I2C1 => (APB1, 21),
+    I2C2 => (APB1, 22),
 }
 #[cfg(feature = "i2c3")]
 bus! {
-    I2C3 => (apb1enr, apb1lpenr, apb1rstr, 23),
+    I2C3 => (APB1, 23),
 }
 #[cfg(feature = "fmpi2c1")]
 bus! {
-    FMPI2C1 => (apb1enr, apb1lpenr, apb1rstr, 24),
+    FMPI2C1 => (APB1, 24),
 }
 
 bus! {
-    USART1 => (apb2enr, apb2lpenr, apb2rstr, 4),
-    USART2 => (apb1enr, apb1lpenr, apb1rstr, 17),
-    USART6 => (apb2enr, apb2lpenr, apb2rstr, 5),
+    USART1 => (APB2, 4),
+    USART2 => (APB1, 17),
+    USART6 => (APB2, 5),
 }
 #[cfg(feature = "usart3")]
 bus! {
-    USART3 => (apb1enr, apb1lpenr, apb1rstr, 18),
+    USART3 => (APB1, 18),
 }
 
 #[cfg(any(feature = "uart4", feature = "uart5"))]
 bus! {
-    UART4 => (apb1enr, apb1lpenr, apb1rstr, 19),
-    UART5 => (apb1enr, apb1lpenr, apb1rstr, 20),
+    UART4 => (APB1, 19),
+    UART5 => (APB1, 20),
 }
 
 #[cfg(any(feature = "uart7", feature = "uart8"))]
 bus! {
-    UART7 => (apb1enr, apb1lpenr, apb1rstr, 30),
-    UART8 => (apb1enr, apb1lpenr, apb1rstr, 31),
+    UART7 => (APB1, 30),
+    UART8 => (APB1, 31),
 }
 #[cfg(any(feature = "uart9", feature = "uart10"))]
 bus! {
-    UART9 => (apb2enr, apb2lpenr, apb2rstr, 6),
-    UART10 => (apb2enr, apb2lpenr, apb2rstr, 7),
+    UART9 => (APB2, 6),
+    UART10 => (APB2, 7),
 }
 
 #[cfg(any(feature = "can1", feature = "can2"))]
 bus! {
-    CAN1 => (apb1enr, apb1lpenr, apb1rstr, 25),
-    CAN2 => (apb1enr, apb1lpenr, apb1rstr, 26),
+    CAN1 => (APB1, 25),
+    CAN2 => (APB1, 26),
 }
 #[cfg(feature = "dac")]
 bus! {
-    DAC => (apb1enr, apb1lpenr, apb1rstr, 29),
+    DAC => (APB1, 29),
 }
 
 bus! {
-    SYSCFG => (apb2enr, apb2lpenr, apb2rstr, 14),
+    SYSCFG => (APB2, 14),
 }
 
 bus! {
-    ADC1 => (apb2enr, apb2lpenr, apb2rstr, 8),
+    ADC1 => (APB2, 8),
 }
 
 #[cfg(feature = "adc2")]
-impl private::Sealed for crate::pac::ADC2 {}
+impl Sealed for crate::pac::ADC2 {}
 #[cfg(feature = "adc2")]
-bus_enable!(ADC2 => (apb2enr, 9));
+impl RccBus for crate::pac::ADC2 {
+    type Bus = APB2;
+}
 #[cfg(feature = "adc2")]
-bus_lpenable!(ADC2 => (apb2lpenr, 9));
+bus_enable!(ADC2 => (APB2, 9));
 #[cfg(feature = "adc2")]
-bus_reset!(ADC2 => (apb2rstr, 8));
+bus_lpenable!(ADC2 => (APB2, 9));
+#[cfg(feature = "adc2")]
+bus_reset!(ADC2 => (APB2, 8));
 
 #[cfg(feature = "adc3")]
-impl private::Sealed for crate::pac::ADC3 {}
+impl Sealed for crate::pac::ADC3 {}
 #[cfg(feature = "adc3")]
-bus_enable!(ADC3 => (apb2enr, 10));
+impl RccBus for crate::pac::ADC3 {
+    type Bus = APB2;
+}
 #[cfg(feature = "adc3")]
-bus_lpenable!(ADC3 => (apb2lpenr, 10));
+bus_enable!(ADC3 => (APB2, 10));
 #[cfg(feature = "adc3")]
-bus_reset!(ADC3 => (apb2rstr, 8));
+bus_lpenable!(ADC3 => (APB2, 10));
+#[cfg(feature = "adc3")]
+bus_reset!(ADC3 => (APB2, 8));
 
 #[cfg(feature = "sdio")]
 bus! {
-    SDIO => (apb2enr, apb2lpenr, apb2rstr, 11),
+    SDIO => (APB2, 11),
 }
 
 bus! {
-    TIM1 => (apb2enr, apb2lpenr, apb2rstr, 0),
-    TIM5 => (apb1enr, apb1lpenr, apb1rstr, 3),
-    TIM9 => (apb2enr, apb2lpenr, apb2rstr, 16),
-    TIM11 => (apb2enr, apb2lpenr, apb2rstr, 18),
+    TIM1 => (APB2, 0),
+    TIM5 => (APB1, 3),
+    TIM9 => (APB2, 16),
+    TIM11 => (APB2, 18),
 }
 
 #[cfg(any(
@@ -272,10 +287,10 @@ bus! {
     feature = "stm32f479"
 ))]
 bus! {
-    TIM2 => (apb1enr, apb1lpenr, apb1rstr, 0),
-    TIM3 => (apb1enr, apb1lpenr, apb1rstr, 1),
-    TIM4 => (apb1enr, apb1lpenr, apb1rstr, 2),
-    TIM10 => (apb2enr, apb2lpenr, apb2rstr, 17),
+    TIM2 => (APB1, 0),
+    TIM3 => (APB1, 1),
+    TIM4 => (APB1, 2),
+    TIM10 => (APB2, 17),
 }
 
 #[cfg(any(
@@ -296,7 +311,7 @@ bus! {
     feature = "stm32f479"
 ))]
 bus! {
-    TIM6 => (apb1enr, apb1lpenr, apb1rstr, 4),
+    TIM6 => (APB1, 4),
 }
 
 #[cfg(any(
@@ -316,9 +331,9 @@ bus! {
     feature = "stm32f479"
 ))]
 bus! {
-    TIM7 => (apb1enr, apb1lpenr, apb1rstr, 5),
-    TIM8 => (apb2enr, apb2lpenr, apb2rstr, 1),
-    TIM12 => (apb1enr, apb1lpenr, apb1rstr, 6),
-    TIM13 => (apb1enr, apb1lpenr, apb1rstr, 7),
-    TIM14 => (apb1enr, apb1lpenr, apb1rstr, 8),
+    TIM7 => (APB1, 5),
+    TIM8 => (APB2, 1),
+    TIM12 => (APB1, 6),
+    TIM13 => (APB1, 7),
+    TIM14 => (APB1, 8),
 }